listos/ticketeck/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
pandas==2.1.3
requests==2.31.0
lxml==4.9.3
psycopg2-binary==2.9.9
pytz==2023.3
python-dateutil==2.8.2